Draw circle
I managed to draw a circle using shaders defined in the code, but
the goal was to use shaders in local files.

This goal was what took me down the build tool rabbit hole to figure
out how to cross compile from wsl for Win32.  I think figuring out the
details there will allow me to read in shader files instead of
having the shaders hard coded within the source.